# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateOrganisationInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  handlers.CreateProductInput:
    properties:
      name:
        type: string
      organisation_id:
        type: integer
    required:
    - name
    - organisation_id
    type: object
  handlers.CreateRepositoryInput:
    properties:
      name:
        type: string
      product_id:
        type: integer
      repo_url:
        type: string
    required:
    - name
    - product_id
    - repo_url
    type: object
...
  models.Component:
    properties:
      compliance_reason:
        type: string
      compliance_status:
        description: allowed, denied, unknown
        type: string
      id:
        type: integer
      license:
        type: string
      name:
        type: string
      purl:
        type: string
      sbom_id:
        type: integer
      type:
        type: string
      version:
        type: string
      vulnerabilities:
        items:
          $ref: '#/definitions/models.Vulnerability'
        type: array
    type: object
  models.Organisation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/models.Policy'
        type: array
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updated_at:
        type: string
    type: object
  models.Policy:
    properties:
      allowed_licenses:
        description: Comma-separated
        type: string
      created_at:
        type: string
      disallowed_deps:
        description: Comma-separated (partial name match)
        type: string
      disallowed_licenses:
        description: Comma-separated
        type: string
      id:
        type: integer
      name:
        type: string
      organisation_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Product:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      organisation_id:
        type: integer
      policy:
        $ref: '#/definitions/models.Policy'
      policy_id:
        type: integer
      repositories:
        items:
          $ref: '#/definitions/models.Repository'
        type: array
      updated_at:
        type: string
    type: object
  models.Repository:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_scan:
        type: string
      name:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      repo_url:
        type: string
      sboms:
        items:
          $ref: '#/definitions/models.SBOM'
        type: array
      status:
        description: pending, scanning, completed, failed
        type: string
      updated_at:
        type: string
    type: object
  models.SBOM:
    properties:
      components:
        items:
          $ref: '#/definitions/models.Component'
        type: array
      created_at:
        type: string
      file_name:
        type: string
      format:
        description: CycloneDX, SPDX, Syft
        type: string
      id:
        type: integer
      repository_id:
        type: integer
      updated_at:
        type: string
      version:
        type: string
    type: object
  models.Vulnerability:
    properties:
      created_at:
        type: string
      cvss:
        type: string
      details:
        type: string
      id:
        type: integer
      osv_id:
        type: string
      severity:
        description: CRITICAL, HIGH, MEDIUM, LOW
        type: string
      summary:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Scapegoat SBOM and Vulnerability Management API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scapegoat API
  version: "1.0"
paths:
...
  /organisations:
    get:
      description: Get a list of all organisations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organisation'
            type: array
      summary: List organisations
      tags:
      - organisations
    post:
      consumes:
      - application/json
      description: Create a new organisation with the given name
      parameters:
      - description: Organisation object
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrganisationInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organisation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new organisation
      tags:
      - organisations
  /organisations/{id}:
    delete:
      description: Delete an organisation by ID
      parameters:
      - description: Organisation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an organisation
      tags:
      - organisations
    put:
      consumes:
      - application/json
      description: Update the name of an existing organisation
      parameters:
      - description: Organisation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organisation object
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrganisationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organisation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an organisation
      tags:
      - organisations
...
  /products:
    get:
      description: Get a list of products, optionally filtered by organisation
      parameters:
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product under an organisation
      parameters:
      - description: Product object
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new product
      tags:
      - products
...
  /repositories:
    get:
      description: Get a list of repositories, optionally filtered by product or organisation
      parameters:
      - description: Product ID to filter by
        in: query
        name: product_id
        type: integer
      - description: Organisation ID to filter by
        in: query
        name: organisation_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Repository'
            type: array
      summary: List repositories
      tags:
      - repositories
    post:
      consumes:
      - application/json
      description: Create a new repository under a product
      parameters:
      - description: Repository object
        in: body
        name: repository
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRepositoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Repository'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new repository
      tags:
      - repositories
  /repositories/{id}:
    get:
      description: Get a repository by ID
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Repository'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a repository
      tags:
      - repositories
  /repositories/{id}/sbom:
    get:
      description: Get the most recent SBOM (with components and vulnerabilities)
        for a repository
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SBOM'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get repository SBOM
      tags:
      - repositories
  /repositories/{id}/scan:
    post:
      description: Trigger a new SBOM scan for the given repository
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "222":
          description: ""
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Trigger a repository scan
      tags:
      - repositories
  /repositories/{id}/scans:
    get:
      description: Get a history of SBOM scans for a repository
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SBOM'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List repository scans
      tags:
      - repositories
...
swagger: "2.0"